Python_App/requirements.txt
anndata==0.8.0
gseapy==0.10.8
klembord==0.3.0
leidenalg==0.8.9
louvain==0.7.1
matplotlib==3.5.1
mnnpy==0.1.9.5
numpy==1.21.0
pandas==1.4.2
Pillow==9.1.1
python_igraph==0.9.11
scanorama==1.7.2
scanpy==1.7.2
scipy==1.8.0
scrublet==0.2.3
seaborn==0.11.2
